fix(billing): charge fractional prices, and print what was actually charged
A line priced 666.66 was rung up as zero. `OrderFoodModel.price` was an `int`
parsed with `parseInt`, whose String branch is `int.tryParse(value) ?? 0` —
and `int.tryParse("666.66")` is null. The replica canonicalises numerics to
strings, so every fractional price took that branch. Prices are fractional in
the ordinary course of business: `goods.price` is `cost_price * (1 + markup%)`,
so any markup that is not a round divisor of the cost produces cents.

The customer was undercharged, and then the payment was lost. `PayOrderBill`
sums the exact decimals and rounds once, and settles only if what was collected
is within 1 so'm. Order 62879273 asked for 76933 against a server figure of
77600 — short by exactly the missing line plus its service — so the API
answered 400, and a 400 is a verdict the outbox quarantines. The regression
test reconstructs that bill.

Price is a double end-to-end now, summed and rounded once, the same arithmetic
the backend does.

The receipts were wrong in three other ways, all found by the same pass:

* Closing a check printed no service fee. `buildFromDetail` re-derived it from
  the order row, and on that path the row is the *unpaid* bill — the backend
  fills `service_amount` when it settles — so it was still 0. Reprinting the
  same bill from the archive, a settled row, showed the fee. The payment screen
  now hands the receipt the `OrderTotals` it charged, so ИТОГО is the amount
  taken rather than re-added from parts, and the preview modal reads the same
  figures it prints.
* Service was charged on the table charge in the waiter's close receipt, which
  the venue does not do and `OrderTotals` never did. The paper asked for more
  than the till took, and disagreed with the cashier's receipt for the same
  bill.
* An archive reprint never subtracted the settled discount, so the copy a guest
  is most likely to be handed in a dispute showed more than they paid.

+++ lib/features/view/main/domain/entities/order_food_entity.dart
@@ -2,7 +2,11 @@ abstract class OrderFoodEntity {
   final String id;
   final String name;
   final int quantity;
-  final int price;
+
+  /// Unit price, in so'm, **with cents**. Not an int: a markup-derived price
+  /// like 666.66 is ordinary, and truncating it here is how a bill silently
+  /// came up short of what the backend charges.
+  final double price;
   final String comment;
   final String status;
   final DateTime? createdAt;
